Persistency (#4)
* preparation

* Not working yet

* remove Response, add some fix

* add caching moves and persons

* remove runCatching

* remove runCatching

* remove data observer

* add "Change spanCount when rotation"

* remove toolbar

+++ app/src/main/java/com/dsvag/androidacademyproject/ui/moviedetails/CastAdapter.kt
@@ -10,6 +10,7 @@ import coil.load
 import com.dsvag.androidacademyproject.R
 import com.dsvag.androidacademyproject.databinding.RowActorBinding
 import com.dsvag.androidacademyproject.models.credits.Cast
+import com.dsvag.androidacademyproject.models.person.Person
 
 class CastAdapter : RecyclerView.Adapter<CastAdapter.CastViewHolder>() {
 
@@ -25,7 +26,7 @@ class CastAdapter : RecyclerView.Adapter<CastAdapter.CastViewHolder>() {
         holder.bind(castList[position])
 
         holder.itemView.setOnClickListener {
-            val bundle = Bundle().apply { putInt("castId", castList[position].id) }
+            val bundle = Bundle().apply { putLong("castId", castList[position].id) }
             holder.itemView.findNavController()
                 .navigate(R.id.action_movieDetailsFragment_to_creditFragment, bundle)
         }
